% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccor.R
\name{ccor}
\alias{ccor}
\title{Scale-Centered (or 'Cohen', rc) Correlations}
\usage{
ccor(x, smin, smax)
ccor(x, smin = 1, smax = 6) #e.g., standard 1-5 Likert scale
}
\arguments{
\item{x}{matrix to be used for correlations}

\item{smin}{Scale minimum}

\item{smax}{Scale maximum}

\item{margin}{Specify whether correlations are of rows(1) or columns(2) (default is columns)}
}
\value{
The scale-centered correlations of matrix
}
\description{
Estimate correlations between variables where the scale midpoint serves as the reflection point for estimating correlations.

These are equivalent to Cohen's (1969) rc correlations, which are invariant across potentially arbitrary item- or variable-reversals.
}
