% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionTest.R
\name{conditionTest}
\alias{conditionTest}
\title{Test of Condition Differences across Situation-Action (s,i) Pairs}
\usage{
ecov(data,fX,fY,dXdY="cov")
}
\arguments{
\item{PSIdata.cond}{PSI-structured datafile, with "condition" as the final}

\item{hiCond}{exact name of 'high' condition (to be scored as '1')}

\item{loCond}{exact name of 'low' condition (to be scored as '-1')}

\item{locEE}{location of expected outcome ratings to test for condition
associations - usually something like \code{c(4:8)},
as first three columns should be \code{p,s,i}}
}
\value{
Correlations and p-Values linking expected outcome
ratings to high/low condition levels
}
\description{
Test whether a (dichotomous) condition produces mean-level
differences in how people rate situation-action expected outcomes
}
\details{
This assumes a very regimented c("p","s","i",...) data structure.
Additionally, the code is currently only for dichotomous condition levels.  If a
condition has more than 2 factor levels, then a different function will have to be
employed (perhaps using 'aov' routine)
}
