% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecov.lmh.R
\name{ecov.lmh}
\alias{ecov.lmh}
\title{Expected-Outcome Covariance Matrices for Low, Med, High Subgroups on a Variable of Interest}
\usage{
ecov.lmh(PSI, pVars, voi, p = "p", cuts = "z", per = F)
}
\arguments{
\item{PSI}{the Person-Situation-Action expected effects relevant to the analysis}

\item{pVars}{file containing person variables}

\item{voi}{(person-)variable of interest (must provide in quotations)}

\item{p}{the variable containing the 'person' identifier (defaults to \code{"p"})}

\item{cuts}{detail how groups will be created; default to \code{z} to
indicate \code{zgroups} will be formed}
}
\value{
Effect covariance matrices (and sample sizes) for \code{zL, zM, zH}
(low, medium, and high z-scores on the person variable of interest)
subgroups
}
\description{
Estimate how expected outcomes covary for 'low', 'medium' and 'high'
levels of a 'person variable of interest' (\code{pvoi}).

Note that this creates 'low' 'medium' and 'high' subgroups through
the strategy employed by the \code{zgroups} function, which attempts
to approximate how observations are weighted when calculating a
correlation coefficient.
}
