% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecov.R
\name{ecov}
\alias{ecov}
\title{Effect Covariance Matrix}
\usage{
ecov(data,fX,fY,dXdY="cov")
}
\arguments{
\item{data}{effect covariance elements (should be generated from \code{ecov_long})}

\item{fX}{Variable containing the X expected outcome feature names}

\item{fY}{Variable containing the Y expected outcome feature names}

\item{dXdY}{Variable containing the product of X and Y expected effects}

\item{order}{specification of the order you would like the variables to be
ordered within the output (defaults to \code{NA} if none given)}
}
\value{
Expected effect covariance (and correlation) matrix; & associated sample sizes
}
\description{
Creates the effect covariance matrix from the long-format
P*S*I effect file (usually estimated by function \code{ecov_long})
}
\details{
This function utilizes functions from \code{plyr} and \code{tidyr}

This will also place features in alphabetical order by default
- which is almost always suboptimal.
This can be altered by specifying an \code{order}, in the form of a
vector with the names of the variables in the preferred order.
}
