% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectiveN.R
\name{effectiveN}
\alias{effectiveN}
\title{Effective N from Weighted Data}
\usage{
effectiveN(vois, id = F)
}
\arguments{
\item{vois}{variables of interest}

\item{id}{is there an identifying variable? Defaults to \code{TRUE}}
}
\value{
effective N for different variables
}
\description{
This will return a list of the effective number of items (effNitems) and
effective number of markers for X (effNXitems) for the different VOIs
contained in specified S matrix.

Note that variables should be on a -1 to 1 scale, where values near -1 or 1
indicate highly loading items, and items near 0 indicate items that should
be weighted negligibly.

(**This function currently has issues when there are missing values...**)
}
