% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expEffects.voi.R
\name{expEffects.voi}
\alias{expEffects.voi}
\title{Expected Effects of Variable-of-Interest}
\usage{
expEffects.voi(ESJT, voiset, voi)
}
\arguments{
\item{ESJT}{Expected effect data}

\item{voiset}{situational variable of interest}

\item{voi}{name of the variable of interest}
}
\value{
expected effects for situational VOI
}
\description{
Give the average effects associated with this type of situation specified as the situational variable of interest (s_voi)

Note: this function is currently defunct; it doesn't handle 'i' feature appropriately
I believe same functionality exists through combination
of \code{siWeightedData} and \code{ecov_long} and \code{ecov}
}
