% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expOutcomes.R
\name{expOutcomes}
\alias{expOutcomes}
\title{Expected Outcomes Matrix}
\usage{
expOutcomes(s, ff)
}
\arguments{
\item{s}{situation vector or matrix, or 'situation states at a given time'}

\item{ff}{field matrix, or 'field at a given time'}
}
\value{
expected situational states for next 'step' (increment of time)
}
\description{
Estimate expected outcomes of current situation
given current field
}
\details{
Input \code{s} can be either a single situation or a range of
possible situations, with the situational states for each given on a separate row
}
