% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medXMY.R
\name{medXMY}
\alias{medXMY}
\title{Run Tests of Normative Simple Mediation Pathways of Total X->Y Relationship (no Moderators)}
\usage{
medXMY(data, X, Y, jSet, all = F)
}
\arguments{
\item{data}{dataset in a PSI long format}

\item{X}{the initiating variable}

\item{Y}{the dependent variable (often "likelihood" - but not necessarily)}

\item{jSet}{location of the set of variables that will be explored as moderated by Z}

\item{all}{include the full lavaan results for each variable (default to \code{F})}
}
\value{
Will return the normative forces along the mediational pathway, separately
for each potential mediator.

Will also return a summary that can be used to quickly glance the estimates of indirect
pathway and its statistical significance.
}
\description{
This will run and extract single-mediator mediation models
('Model 1' in Wood, Harms, & Cho, 2023) for all measured potential mediators of the
[X->Y] association.
}
\details{
This adjusts for degrees of freedom by using 'p' as a clustering variable in lavaan

Note that this model is referred to as the 'Minimal EV Model' in Wood et al., 2023,
and will produce a simple mediational model that is equivalent to the common
\code{c = a*b + c} formula found in mediational models, but using a different notation -
specifically \code{B0_YX = B1_MX*B1_YM + B1_YX}.

The B0_YX is returned as a bit of a quality check.  If the estimates, p values, etc, are not
isn't the same across all cases, then this indicates that you did not get rid
of cases with missing values, which is generally recommended.
}
