% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modXMY.R
\name{modXMY}
\alias{modXMY}
\title{Run Tests of Moderated-Mediation Pathways of Total X->Y Relationship}
\usage{
modXMY(data, X, Y, Z, jSet, all = F)
}
\arguments{
\item{data}{dataset in a PSI long format}

\item{X}{the initiating variable}

\item{Y}{the dependent variable (often "likelihood" - but not necessarily)}

\item{Z}{moderator variable (it will be standardized)}

\item{jSet}{location of the set of variables that will be explored as moderated by Z}

\item{all}{include the full lavaan results for each variable (default to \code{F})}
}
\value{
Will return the main effects of forces along the mediational pathway, and the z-moderation of each force
}
\description{
This will run and extract single-mediator moderated-mediation models
('Model 1' in Wood, Harms, & Cho, 2023) for all measured potential mediators of the
[X->Y] association.
}
\details{
This adjusts for degrees of freedom by using 'p' as a clustering variable in lavaan
}
