% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parView.R
\name{parView}
\alias{parView}
\title{Provide Rounded Parameter Estimates}
\usage{
parView(fit, split = "est", dec = 3)
}
\arguments{
\item{fit}{standard output from lavaan}

\item{split}{locate where variables will start to be rounded}

\item{dec}{number of variables to round to}
}
\value{
parameter estimates rounded down to desired number of decimal spaces
}
\description{
This will return the parameter estimates from lavaan, but with values
rounded to a desired numer of decimal spaces.
}
