% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnLevels.R
\name{pnLevels}
\alias{pnLevels}
\title{Positive and Negative Levels'}
\usage{
pnLevels(levels)
}
\arguments{
\item{levels}{?the level of the factor? Need to review this some more}
}
\description{
A small modification to qgraph that allows cells to be
colored as a gradient of the level of the variable, on a scale from range
\code{[-1:1]}
}
\details{
?Yeah, this function needs a little review to figure out how it is working within \code{qgraph}
}
