% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdESJT.R
\name{sdESJT}
\alias{sdESJT}
\title{Standard deviation of ESJT ratings}
\usage{
sdESJT(PSIdata, rmvLast = T)
}
\arguments{
\item{rmvLast}{remove the final variable (e.g., if it is Likelihood
ratings provided at different portion of survey)}

\item{ESJT}{ESJT matrix (should have 'p','s','i' as their first three features)}
}
\value{
sdE.p - the percentage of situation-action pairs in which
the respondent showed zero response variability whatsoever.
}
\description{
This will index the standard deviation of the person's ratings across all of the
ESJT scenarios they rated.
}
\details{
It is recommended within Wood, Lowman, & Harms (202x) that
respondents showing values of sd.p below 25\% of the scale maximum be removed from all analyses
}
\examples{
#combine this function and recommended screen into a single statement:
analysiscases <- subset(sdESJT(ESJT), sdE.p > .25)
}
