% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siWeightedData.R
\name{siWeightedData}
\alias{siWeightedData}
\title{Situation-Action Weighted Data}
\usage{
siWeightedData(ESJT, voiset, "voi")
}
\arguments{
\item{ESJT}{ESJT raw data; must be headed with three variables: \code{p, s, i} -
for 'Person', 'Situation', '[Initiated] action', respectively}

\item{voiset}{Set containing the variable of interest; must be headed with ONLY
variables  \code{s, i}.}

\item{voi}{Name of the specific variable within the voiset you wish to weight by
- should be a variable name given in "quotes"}

\item{bySet}{columns you wish to match by (defaults to 'situation & action')}
}
\value{
The expected effects of the action, weighted by the action's situation-diagnosticity
}
\description{
Weight situation-action expected effects by a variable of interest.
Of value for combining multiple situation-action pairs into a single situation,
weighted by the situation-action pair's trait diagnosticity.
(**This function is intended to replace "sweighteddata" soon.)
}
