% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strLine.R
\name{strLine}
\alias{strLine}
\title{Percentage Straight-Lining Behavior Within Actions}
\usage{
strLine(ESJT, rmvLast = T)
}
\arguments{
\item{ESJT}{ESJT matrix (should have 'p','s','i' as their first three features)}

\item{rmvLast}{remove the final variable (e.g., if it is Likelihood
ratings provided at different portion of survey)}
}
\value{
per0.p - the percentage of situation-action pairs in which
the respondent showed zero response variability whatsoever.
}
\description{
This will return the percentage of situation-action pairs for which the
respondent showed zero response variability.
}
\details{
It is recommended within Wood, Lowman, & Harms (202x) that
respondents showing 10% straightlining by this index be removed from all analyses
}
