% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surgery.R
\name{surgery}
\alias{surgery}
\title{Matrix Surgery}
\usage{
surgery(x, y, suppressOverwrite = T)
}
\arguments{
\item{x}{the original matrix is 'old' matrix representing function field model paths}

\item{y}{matrix with non-missing cells you wish to overwrite cells in x}

\item{suppressOverwrite}{do you wish reports of overwritten cells to be suppressed? Defaults to \code{TRUE}}
}
\value{
modified version of \code{x} overwritten with non-missing cells from \code{y}
}
\description{
This function serves as a 'brute force' method for
replacing values in some matrix \code{x} with non-missing values of \code{y}.
}
\details{
Any cell in \code{y} given as a missing value (\code{NA}) will return the original cell in \code{x}

Often, you will NOT want non-missing cells in \code{x} to be overwritten. (This might indicate some problem in the model.)
If so, you can specify \code{suppressOverwrite = F} to print when overwrites have occurred.
}
