% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vshapes and vsizes.R
\name{vshapes}
\alias{vshapes}
\title{Variable Shapes}
\usage{
vshapes(vtype)
}
\arguments{
\item{vtype}{Variable type (short codes)}
}
\description{
This is mainly used to run within the \code{prepField} function
}
