% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vshapes and vsizes.R
\name{vsizes}
\alias{vsizes}
\title{Variable Sizes}
\usage{
vsizes(vtype)
}
\arguments{
\item{vtype}{Variable type (short codes)}
}
\description{
This is mainly used to run within the \code{prepField} function
}
