% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedFF.R
\name{weightedFF}
\alias{weightedFF}
\title{Weight Expectancy Data}
\usage{
sweighteddata(FFdata, voiset, voi)
}
\arguments{
\item{Ex}{ESJT raw data; must be headed with three variables, which should
generally be \code{p, s, i} (for person, situation, and action, respectively)}

\item{voiset}{Set containing the variable of interest}

\item{voi}{Name of the specific variable within the voiset you wish to weight by}

\item{si_ids}{= Name of situation  and  action identifiers, common to both files}
}
\value{
The expected effects of the action, weighted by the action's situation-diagnosticity
}
\description{
In brief: weight situations by a situational and/or action variable of interest.
Mainly of value for eventually combining multiple scenarios into a single score.
For instance: combining diverse ESJT ratings to a single score weighted by the situation's assertiveness-diagnosticity.
}
