% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withinParTable.R
\name{withinParTable}
\alias{withinParTable}
\title{Within-Person N Adjustment for ParTable}
\usage{
withinParTable(fit, data)
}
\arguments{
\item{fit}{standard output from an sem(...) statement in lavaan}

\item{data}{original 'long' format data frame that was used in sem(...) statement}
}
\value{
parTable in which se, z, and p statistics are adjusted to person level
}
\description{
The standard 'parTable' in lavaan does not provide correct degrees of
freedom when working with a deviation-score data file; this corrects that
}
\details{
(This may become outdated as procedures for doing multilevel analyses with field
models become further developed.  It should also substantially underestimate significance, I believe
if 3 or more observations are within-person)
}
