% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgroups.R
\name{zgroups}
\alias{zgroups}
\title{Create High-Medium-Low Z-Score Groups}
\usage{
zgroups(voi)
}
\arguments{
\item{voi}{variable of interest}
}
\value{
weight for low, medium, and high z-score groups
}
\description{
This will return three variables detailing the weight that each individual should
receive for subset analyses to look at how relatively "high" or "low" scorers
on the measure
}
\details{
Note that the sum of the three groups will equal the total sample
of non-missing values.  Can be used in combination  with \code{effectiveN}.
}
