% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zPSI.R
\name{zPSI}
\alias{zPSI}
\title{Scale-Center-Standardized PSI Ratings}
\usage{
zPSI(PSIdata)
}
\arguments{
\item{PSIdata}{original PSI data (should have 'p','s','i' for first three columns)}
}
\value{
Standardized scale-centered scores.

Original matrix but with data now equating level of variability for each participant
}
\description{
Transforms ESJT ratings into a z-score-like metric indicating
the number of 'standard-deviations-from-zero' the person's original rating of the item was.
}
\details{
This syntax implicitly assumes that 0 is a meaningful point on the scale.
The original data collection and the prior processing of the data should reflect this.

* The scale-centered scores do NOT have the usual interpretation of
[regular] z-scores as being translatable to approximate probabilities assuming
a normal distribution of responses.

* If the person gives the same score to all items, all \code{zcx}
will equal \code{1}.
}
