% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devPSI.R
\name{devPSI}
\alias{devPSI}
\title{Deviation Action Scoring of PSI matrix}
\usage{
devPSI(PSIdata, locX = c(4:ncol(PSIdata)))
}
\arguments{
\item{PSIdata}{PSI-structured dataframe (should have 'p','s','i' as their first three features)}

\item{locX}{location of expected outcome ratings (note: default generously assumes
the prescribed 'c(p,s,i,...,Likelihood)' data structure)
ratings provided at different portion of survey)}
}
\value{
deviated ratings of effects of action i as rated by p within situation s
from the person's average rating of actions in that situation.
}
\description{
Will estimate how p's rating of outcomes of each action deviate within situation s
deviated from their average rating of outcomes of different actions in that situation
}
\details{
This essentially forces 'deviation' scoring.

Note that when nI = 2, the deviation scores will be reflections of one another.
In which case, keeping only the maximum i (which should be i=1)  can be useful
to help avoid artificial inflation of sample size
(although other strategies, like multilevel modeling in lavaan,
may be preferable to address this issue).

If we want the difference of the two actions (rather than deviation scoring),
we can pair the devPSI function with diffPSI.
}
\examples{
(develop some example code)
}
