% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecov.R
\name{ecov}
\alias{ecov}
\title{Effect Covariance Matrix}
\usage{
ecov(data,fA,fB,dAdB="cov")
}
\arguments{
\item{data}{Input matrix}

\item{fA}{Variable containing the X expected outcome feature names}

\item{fB}{Variable containing the Y expected outcome feature names}

\item{dAdB}{Variable containing the product of A and B expected effects (note: should be in quotation marks, as in "r" or "cov")}
}
\value{
Expected effect covariance matrix
}
\description{
Creates the effect covariance matrix from the long-format P*S effect file (usually estimated by function \code{ecov_big})
}
\details{
Note: this function depends on having loaded \code{reshape2} package
}
