% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecov_big.R
\name{ecov_big}
\alias{ecov_big}
\title{Estimate Basic Covariance Elements}
\usage{
ecov_big(wFFdata)
}
\arguments{
\item{wFFdata}{weighted functional field data}
}
\value{
Effect covariance matrix
}
\description{
Compute all possible products of effects for all
features from weighted ESJT data, \code{wFFdata}.
     Note that the average of the \code{dAdB} elements will ultimately
equal the covariances forming the expectancy matrix used to
estimate functional field models.

Note: data can be weighted using \code{sweighteddata} function
}
\details{
Usage notes: FFdata should have format of \code{P_ID}, \code{S_ID}, \code{Did_A}, as first three columns,
followed by effect ratings, and end in \code{Likelihood}
}
