% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecov_long.R
\name{ecov_long}
\alias{ecov_long}
\title{Estimate Action Expected-Outcome Covariance Elements}
\usage{
ecov_long(wFFdata, P_ID = "P_ID", S_ID = "S_ID", A_ID = "A_ID")
}
\arguments{
\item{wFFdata}{weighted functional field data}

\item{P_ID}{variable identifying 'person' (or respondent) - give in "quotemarks"}

\item{S_ID}{variable identifying 'situation' (or scenario) - give in "quotemarks"}

\item{A_ID}{variable identifying 'action' (or response) - give in "quotemarks"}
}
\value{
Effect covariance matrix
}
\description{
Compute all possible products of effects for all
features from weighted ESJT data, \code{wFFdata}.
     The average of the \code{dAdB} elements will ultimately
equal the covariances forming the expectancy matrix used to
estimate functional field models.
     Note #1: this function a more generalized version of function \code{ecov_big}, which 
provides flexibility of allowing multiple actions per scenario
     Note #2: this function assumes that the features have a "meaningful zero point."
That is a \code{0} should mean "no effect", and negative values should mean 
"negative effect", and positive values should mean "positive effect", with greater
deviations from zero meaning "stronger effect."  If data is not prepared this way,
the results are not very likely to be meaningful.

Note: data can be weighted using \code{sweighteddata} function
}
\details{
Usage notes: FFdata should have format of \code{P_ID}, \code{S_ID}, \code{A_ID}, as first three columns,
followed by effect ratings, and end in \code{Likelihood}
}
