% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expOutcomesFF.R
\name{expOutcomesFF}
\alias{expOutcomesFF}
\title{Expected Outcomes from Situation Interacting with Functional Field}
\usage{
expOutcomesFF(s, i = 0, ffield = ff, plan = DoNothing)
}
\arguments{
\item{s}{situation vector: 'situation states at a given time'}

\item{i}{action definition - specify the FEATURES you wish to set to value of 1 ("TRUE" or "present")}

\item{ffield}{functional - contains the 'belief system' or 'model' of how force levels are produced}

\item{plan}{specify forces to be added to the functional field}

\item{xS}{extra situational features that are NOT graphed that can add forces to the field}
}
\value{
'Situation log' - detailing \code{sit} level of situational forces,
 \code{dit} new disturbances, and \code{fit} structure of the field at time \code{t}
}
\description{
Estimate expected outcomes of current situation
given a particular functional field (which details how
situational factors determine the levels of forces in the environment)
}
