% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strLine.R
\name{strLine}
\alias{strLine}
\title{Percentage Straight-Lining Behavior Within Actions}
\usage{
strLine(ESJT, locX = c(4:(ncol(ESJT) - 1)))
}
\arguments{
\item{ESJT}{ESJT matrix (should have 'p','s','i' as their first three features)}

\item{locX}{location of expected outcome ratings (note: default generously assumes
the prescribed 'c(p,s,i,...,Likelihood)' data structure)
ratings provided at different portion of survey)}
}
\value{
per0.p - the percentage of situation-action pairs in which
the respondent showed zero response variability whatsoever.
}
\description{
This will return the percentage of situation-action pairs for which the
respondent showed zero response variability.
}
\details{
It is recommended within that respondents showing
10\% straightlining by this index be removed from all analyses,
however less stringent cutoffs (e.g., 25\%) might be employed when
respondents rate very few features (e.g., 5 or less) or if all features
are desirable, both of which can make straight-lining behavior less suspect
}
\examples{
#combine this function and recommended screen into a single statement:
analysiscases <- subset(strLine(ESJT), per0.p > .10)
}
